Adding one level of factorization for generation windows Saving the last used folders paths Updating unity assets when generation is done
