+++ Unity3D/Assets/RosSharp/Scripts/RosBridgeClient/Editor/MessageGeneration/PackageAutoGenEditorWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99b523215e0722744b6c01c139aaf07a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
