+++ Unity3D/Assets/RosSharp/Scripts/RosBridgeClient/Editor/MessageGeneration/SingleMsgAutoGenEditorWindow.cs
@@ -13,106 +13,38 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-using System.IO;
-
 using System.Collections.Generic;
-
-using UnityEngine;
 using UnityEditor;
+using UnityEngine;
 
 namespace RosSharp.RosBridgeClient.MessageGeneration
 {
-    public class SingleMsgAutoGenEditorWindow : EditorWindow
+    public class SingleMsgAutoGenEditorWindow : SingleAutoGenEditorWindow
     {
-        private string inFilePath = "";
-        private string outFilePath = Path.Combine(System.Environment.CurrentDirectory, "Assets", "RosSharpMessages");
-        private string rosPackageName = "";
+        protected override string GenerationType
+        {
+            get { return "message"; }
+        }
+
+        protected override string FileExtension
+        {
+            get { return "msg"; }
+        }
+
 
         [MenuItem("RosBridgeClient/Auto Generate Messages/Single Message...", false, 0)]
-        private static void OpenWindow() {
+        private static void OpenWindow()
+        {
             SingleMsgAutoGenEditorWindow window = GetWindow<SingleMsgAutoGenEditorWindow>(false, "Message Auto Generation", true);
             window.minSize = new Vector2(750, 100);
             window.maxSize = new Vector2(750, 100);
             window.Show();
         }
 
-        private void OnGUI() {
-            GUILayout.Label("Single message auto generation", EditorStyles.boldLabel);
-
-            EditorGUILayout.BeginHorizontal();
-            inFilePath = EditorGUILayout.TextField("Input File Path", inFilePath);
-            if (GUILayout.Button("Browse File...", GUILayout.Width(120))) {
-                inFilePath = EditorUtility.OpenFilePanel("Select Message File...", "", "msg");
-                if (!inFilePath.Equals("")) {
-                    string[] directoryLevels = inFilePath.Split('/');
-                    rosPackageName = directoryLevels[directoryLevels.Length - 3];
-                }
-            }
-            EditorGUILayout.EndHorizontal();
-
-            rosPackageName = EditorGUILayout.TextField("ROS Package Name:", rosPackageName);
-
-            EditorGUILayout.BeginHorizontal();
-            outFilePath = EditorGUILayout.TextField("Output File Location", outFilePath);
-            if (GUILayout.Button("Select Folder...", GUILayout.Width(120)))
-            {
-                outFilePath = EditorUtility.OpenFolderPanel("Select Folder...", "", "");
-            }
-            EditorGUILayout.EndHorizontal();
-
-            if (GUILayout.Button("GENERATE!")) {
-                if (inFilePath.Equals(""))
-                {
-                    EditorUtility.DisplayDialog(
-                        title: "Error", 
-                        message: "Empty input file path!\nPlease specify input file", 
-                        ok: "Bricks without straw");
-                }
-                else
-                {
-                    try
-                    {
-                        List<string> warnings = MessageAutoGen.GenerateSingleMessage(inFilePath, outFilePath, rosPackageName);
-                        if (warnings.Count == 0)
-                        {
-                            EditorUtility.DisplayDialog(
-                                title: "Code Generation Complete", 
-                                message: "Output at: " + outFilePath, 
-                                ok: "Thank you!");
-                        }
-                        else {
-                            foreach (string w in warnings) {
-                                Debug.LogWarning(w);
-                            }
-                            EditorUtility.DisplayDialog(
-                                title: "Code Generation Complete", 
-                                message: "Output at: " + outFilePath + "\nYou have " + warnings.Count + " warning(s)", 
-                                ok: "I like to live dangerously");
-                        }
-                    }
-                    catch (MessageTokenizerException e)
-                    {
-                        Debug.LogError(e.ToString() + e.Message);
-                        EditorUtility.DisplayDialog(
-                            title: "Message Tokenizer Exception", 
-                            message: e.Message, 
-                            ok: "Wait. That's illegal");
-                    }
-                    catch (MessageParserException e)
-                    {
-                        Debug.LogError(e.ToString() + e.Message);
-                        EditorUtility.DisplayDialog(
-                            title: "Message Parser Exception", 
-                            message: e.Message, 
-                            ok: "Sorry but you can't ignore errors.");
-                    }
-                }
-            }       
-        }
-
-        private void OnInspectorUpdate()
+        protected override List<string> Generate(string inPath, string outPath, string rosPackageName = "")
         {
-            Repaint();
+            return MessageAutoGen.GenerateSingleMessage(inPath, outPath, rosPackageName);
         }
+
     }
 }
